requirements.txt
pandas
spotipy
boto3
python-dotenv